refactor(settings): move SettingsCache from auth to settings package
Define SettingProvider interface in settings package, breaking the
coupling where 6+ non-auth packages imported auth just for SettingsCache.

- New: settings.SettingsCache with SettingProvider interface
- auth.SystemSettingsService implements SettingProvider via AsProvider()
- auth.SettingsCache kept as type alias for backward compat
- 10+ files updated to use settings.SettingsCache directly
- Middleware, email, API handlers no longer depend on auth for caching
